refactor(socket): move socket type declarations to dedicated file
The socket type declarations were moved from socket.ts to socket.d.ts to better organize type definitions and follow TypeScript best practices for type declaration files.
