+++ src/types/socket.d.ts
@@ -1,0 +1,10 @@
+import { Socket } from "socket.io";
+
+declare module "socket.io" {
+	interface Socket {
+		userId?: string;
+		userRole?: string;
+		vendorId?: string;
+		isAuthenticated?: boolean;
+	}
+}
